backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# Celery y Redis
celery==5.3.4
redis==5.0.1

# Playwright para automatización
playwright==1.41.0
playwright-stealth==1.0.6

# IA y traducciones (CAMBIO: Sustituimos OpenAI por Google Generative AI)
google-generativeai>=0.8.0

# Peticiones HTTP
httpx>=0.26.0

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0